requirements.txt
streamlit
opencv-python
opencv-python-headless
numpy
Pillow
tensorflow
spacy
easyocr
transformers
pyspellchecker
pyrebase4
flask
flask-cors